erlang21: Re-import erlang version 21.3.
This is a copy of the previous lang/erlang of that release, with a few bits
tidied up, and configured to live alongside the newer lang/erlang package by
using a versioned library directory.

This version is required for at least databases/couchdb, and may be required
for other software that is incompatible with OTP 23.