nghttp2: updated to 1.43.0
v1.43.0:

doc

Documentations are now built with Sphinx 3.3.0 or later.

python

The python binding now requires Python 3.
All python scripts for nghttp2 development are translated to Python 3 compatible.

nghttpx

This release fixes a potential memory issue that a memory pool gets cleared while it is still in use.
ECDSA certificate is now chosen when compatible signature algorithm is available.
This release adds a workaround to include ‘:’ in backend pattern.