tex-pgf{,-doc}: update to 3.1.8b
3.1.8b
Hotfix for handling of TeX conditionals on a path.  We can't forward
\relax and frozen \relax through the parser because there is existing
code that relies on this.

The recommendation is to use expandable conditionals where possible.

3.1.8a
Hotfix for the new topaths handling.  One instance did not yet properly
preserve relative coordinates.

3.1.8
If a topath is bent by any of the in=, out=, bend=, etc. options, a
Bezier curve is constructed in the background.  To infer the positions
of the control points the start and end coordinate are converted to
absolute coordinates. However, this has the effect that subsequent
points on the path think that the endpoint of the topath was absolute
which can lead to counter-intuitive path construction, e.g.
```latex
\draw (2,0) to[out=0,in=180] +(1,0) -- ++(0,-1) -- +(1,0);
```
If old code relies on this behavior, this drawing will silently break!
Please open an issue if you rely on this.

This release introduces a fix for path handling which concerns
expansion of tokens on the path in particular with respect to
conditional.  Previously when the expansion of a conditonal resulted in
a frozen \relax the parser would just give up.  Now the parser will
skip over the frozen \relax and continue to expand tokens.  Whether
this will result in a meaningful expansion is up to the user.

This release also includes other bug fixes.